# packages/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: CineWeave API
  version: 0.1.0
  description: API contract seed for the Phase 1 CineWeave control plane.
servers:
  - url: http://localhost:8080
paths:
...
  /api/projects/{projectId}/scripts/{scriptId}:
    get:
      operationId: getScript
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/ScriptId"
      responses:
        "200":
          description: Script detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScriptEnvelope"
    patch:
      operationId: updateScript
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/ScriptId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateScriptRequest"
      responses:
        "200":
          description: Updated script
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScriptEnvelope"
...
  /api/projects/{projectId}/scripts/{scriptId}/versions/{versionId}/parse-scenes:
    post:
      operationId: parseScriptScenes
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/ScriptId"
        - name: versionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParseScriptScenesRequest"
      responses:
        "200":
          description: Parsed structured script scenes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParseScriptScenesResponseEnvelope"
  /api/projects/{projectId}/scripts/{scriptId}/scenes:
    get:
      operationId: listScriptScenes
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/ScriptId"
        - name: scriptVersionId
          in: query
          schema:
            type: string
            format: uuid
        - name: reviewStatus
          in: query
          schema:
            $ref: "#/components/schemas/ReviewStatus"
      responses:
        "200":
          description: Structured script scenes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScriptSceneListEnvelope"
  /api/projects/{projectId}/script-scenes/{sceneId}:
    get:
      operationId: getScriptScene
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - name: sceneId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Structured script scene detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScriptSceneEnvelope"
    patch:
      operationId: updateScriptScene
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - name: sceneId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateScriptSceneRequest"
      responses:
        "200":
          description: Updated structured script scene
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScriptSceneEnvelope"
  /api/projects/{projectId}/script-scenes/{sceneId}/review:
    post:
      operationId: reviewScriptScene
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - name: sceneId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewRequest"
      responses:
        "200":
          description: Updated structured script scene review status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewResponseEnvelope"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    serviceTokenAuth:
      type: http
      scheme: bearer
      bearerFormat: service-token
  parameters:
...
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    ScriptId:
      name: scriptId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
    Envelope:
      type: object
      required: [requestId]
      properties:
        requestId:
          type: string
        data: {}
        meta: {}
...
    Script:
      type: object
      required: [id, organizationId, projectId, title, status, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        sourceId:
          type: string
          format: uuid
        title:
          type: string
        status:
          type: string
          enum: [draft, active, archived]
        currentVersionId:
          type: string
          format: uuid
        currentVersion:
          $ref: "#/components/schemas/ScriptVersion"
        createdBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ScriptVersion:
      type: object
      required: [id, organizationId, projectId, scriptId, version, content, contentFormat, metadata, createdAt]
      properties:
        id:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        scriptId:
          type: string
          format: uuid
        version:
          type: integer
        content:
          type: string
        contentFormat:
          type: string
          enum: [plain_text, markdown]
        sourceType:
          type: string
        promptVersionId:
          type: string
          format: uuid
        promptHash:
          type: string
        metadata:
          type: object
          additionalProperties: true
        createdBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
    ScriptScene:
      type: object
      required: [id, projectId, scriptId, scriptVersionId, sceneIndex, sceneNo, title, characters, scenes, props, sourceEventIds, content, contentFormat, reviewStatus, manualOverride, staleState]
      properties:
        id:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        scriptId:
          type: string
          format: uuid
        scriptVersionId:
          type: string
          format: uuid
        sceneIndex:
          type: integer
        sceneNo:
          type: integer
        title:
          type: string
        summary:
          type: string
        location:
          type: string
        timeOfDay:
          type: string
        atmosphere:
          type: string
        characters:
          type: array
          items:
            type: string
        scenes:
          type: array
          items:
            type: string
        props:
          type: array
          items:
            type: string
        action:
          type: string
        dialogue:
          type: string
        visualGoal:
          type: string
        emotionalTone:
          type: string
        conflict:
          type: string
        outcome:
          type: string
        sourceEventIds:
          type: array
          items:
            type: string
        content:
          type: string
        contentFormat:
          type: string
          enum: [plain_text, markdown]
        reviewStatus:
          $ref: "#/components/schemas/ReviewStatus"
        manualOverride:
          type: boolean
        staleState:
          $ref: "#/components/schemas/StaleState"
        metadata:
          type: object
          additionalProperties: true
        createdBy:
          type: string
          format: uuid
        editedBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        editedAt:
          type: string
          format: date-time
...
    ReviewStatus:
      type: string
      enum: [pending, approved, rejected, needs_edit]
    StaleState:
      type: string
      enum: [fresh, upstream_changed, needs_regeneration]
...
    ReviewRequest:
      type: object
      required: [reviewStatus]
      properties:
        reviewStatus:
          $ref: "#/components/schemas/ReviewStatus"
        note:
          type: string
    ReviewResponse:
      type: object
      required: [id, reviewStatus, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        reviewStatus:
          $ref: "#/components/schemas/ReviewStatus"
        note:
          type: string
        updatedAt:
          type: string
          format: date-time
...
    ReviewResponseEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/ReviewResponse"
...
    ScriptEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/Script"
...
    ScriptSceneEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/ScriptScene"
    ScriptSceneListEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              type: object
              properties:
                items:
                  type: array
                  items:
                    $ref: "#/components/schemas/ScriptScene"
    ParseScriptScenesResponseEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/ParseScriptScenesResponse"
...
    UpdateScriptRequest:
      type: object
      properties:
        sourceId:
          type: string
          format: uuid
        title:
          type: string
        status:
          type: string
          enum: [draft, active, archived]
...
    ParseScriptScenesRequest:
      type: object
      properties:
        force:
          type: boolean
          default: false
    ParseScriptScenesResponse:
      type: object
      required: [scriptId, versionId, sceneCount, scenes]
      properties:
        scriptId:
          type: string
          format: uuid
        versionId:
          type: string
          format: uuid
        sceneCount:
          type: integer
        scenes:
          type: array
          items:
            $ref: "#/components/schemas/ScriptScene"
        providerCallId:
          type: string
          format: uuid
        modelId:
          type: string
    UpdateScriptSceneRequest:
      type: object
      properties:
        title:
          type: string
        summary:
          type: string
        location:
          type: string
        timeOfDay:
          type: string
        atmosphere:
          type: string
        characters:
          type: array
          items:
            type: string
        scenes:
          type: array
          items:
            type: string
        props:
          type: array
          items:
            type: string
        action:
          type: string
        dialogue:
          type: string
        visualGoal:
          type: string
        emotionalTone:
          type: string
        conflict:
          type: string
        outcome:
          type: string
        content:
          type: string
...